requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
